requirements.txt
Flask==2.3.2
gspread==5.12.0
oauth2client==4.1.3
pytesseract==0.3.10
Pillow==9.5.0
google-api-python-client==2.126.0
openai==1.30.1
python-dotenv==1.0.1
